Refactor validations creator to accomodate require_if logic, move types
